Consolidate the parts required to run a Task (#2532).
Ticket #2303 (multiprocessing for CmdLineTask) made it more difficult to
override the methods (e.g., to add additional inputs to the run method)
because it requires making changes in 3 places, including a free function
(because it has to be picklable).

The free function and the method for interpreting the arguments have moved
to a new class (TaskRunner) that has a __call__ method. This new class
provides a common place for changes to be made, so that most of the
changes required are at least in some common place (and in something that
can be subclassed, and not floating free).