#2144: support multiple orthogonal identifiers.
New method to add an argument (add_id_argument) that sets up the
information required to process the command-line.  The results are
available to the user as a Struct with 'idList' and 'refList' elements.
This differs from a previous implementation of this ticket, where a
custom class was used as the carrier; the new implementation allows us to
keep the method for parsing the data identifiers (_castDataIds) and for
converting the data identifiers to data references (_makeDataRefList) as
members of the ArgumentParser, and therefore easily overriden by the user.