Cleaning up in response to Russell's comments.
Using a custom class instead of Struct to hold the identifier data.  This is
minimally different from Struct (it inherits from it), but provides a
convenience when debugging.

Using a custom class instead of Struct to hold the results of the identifier
parsing.  Moved the casting and reference construction there, which makes
overriding the behaviour easier.  add_id_argument takes the Class name,
which allows the override.