Modify CmdLineTask.getTargetList to provide kwargs.
This should make it slightly easier to provide additional arguments to the
Task.run() methods beyond just the data reference.

Russell writes:
"""
[I suggest we] have getTargetList return a tuple: (dataRef, argDict)

This is less general and than your idea, but I prefer it because:
- It is generally better to pass arguments by name. This design encourages that
  behavior for any arguments past dataRef. (See note for why this dataRef is
  special).
- It discourages the idea that the user can override getTargetList to return any
  arbitrary argList (as well as argDict): it would break the default __call__
  because that requires a dataRef in several places.

Note: I usually dislike passing args by position, but for tasks it seems to make
sense to pass a dataRef that way because then the dataRef argument can be named
for the kind of data, e.g. sensorRef, patchRef, etc. However, as soon as a task
takes anything beyond a single dataRef in its run method then TaskRunner must be
overridden to be task-specific, so why not have the remaining arguments called
by name?
"""