TaskRunner: make multiprocessing interruptible (#3039)
multiprocessing.Pool.map isn't interruptible (i.e., Ctrl-C) because
the interrupt kills one of the processes, which then isn't able to
re-join. Using pool.map_async addresses this.  This requires specifying
a timeout, which is provided as a class variable, for the user to easily
modify (an instance variable would be difficult to modify without
subclassing).