Make doReturnResults more robust for CmdLineTask
If a command-line task fails and doReturnResults is True and doRaise is False
then an unhelpful exception occurs instead of returning the results.
This commit fixes the problem by setting the task result to None.
It also adds a unit test for the condition.
The unit test requires an updated obs_test (for its updated TestConfig).