Replace --trace with --logdest in ArgumentParser
Enhanced --logdest to support name=value arguments (like --trace)
and eliminated --trace.
Updated the unit test to test this modified functionality.

The reason for this change is that use of pex_logging Trace and Debug
is deprecated; all such logging should be done with Log.
We are planning to switch from pex_logging to lsst.log and that
will force this change, since lsst.log has no Trace or Debug.
Thus we should have a way to set log level by log name.
That is what --trace does, but --trace does not support
named log levels and log level = - trace level, so using --trace
for --loglevel is confusing. It is also confusing to have two
different arguments that can and should do the same thing.