Allow for exact config name in --show config=glob
Previously, entering an exact (non-glob) config name to
--show config=glob
would return nothing if the glob was an exact config name (i.e.
with no * wildcard as the rightmost character)
e.g.
--show config=*badStatistic* returned config.isr.badStatistic='MEANCLIP'
whereas
--show config=config.isr.badStatistic returns nothing

This happened because the string comparison was being done on the
full config setting, i.e. including the "=VALUE" portion.  This
commit remedies that situation by trimming the "=VALUE" portion
from the string used for matching.