Update unittest to account for doc string printing
We now include printing of documentation strings for config
parameters.  testArgumentParser.py's testShow() function needed
updating to account for this new output.  An additional config Field
item was also added to the testing to ensure all is ok when the
docstring is multiline.

The specific commits that necessitated this changes are:

Config: when printing config fields also include the documentation
https://github.com/lsst/pex_config/commit/9240a2d86a2a0189457681ac598b2f86be6d50f7

and

Use "#" instead of triple quotes for doc printing
https://github.com/lsst/pex_config/commit/cba14cd246b438e783710665b50560ef18ebadc7