Introduce rerun command line flag
The rerun command is functionally similar to the output argument, but the
value supplied will be treated as relative to the rerun directory of the
root input data repository. This eliminates the need for the user to type
the input repository twice. --rerun also supports supplying an optional
output directory which is specified as input:output to the rerun command,
making it easier to pick up previous work and re-target the output in the
next stage of processing

This functionality was pulled from the HSC fork as is, without history.
The history on the HSC was split amongst many individual commits, and
replaying the commits would have many conflicts such that the resulting
history would be more difficult to follow