Remove setting default output in argument parser
The code ported from HSC which introduced the rerun command line switch,
also introduced a default output if no output was otherwise specified. This
output was built using the rerun directory semantics. As introducing a new
default output is beyond the scope of this ticket the functionality is
being removed from this ticket, but will be revisited in DM-4236, where the
new default behavior will be introduced, and the corresponding
documentation updated.