ArgumentParser: require an output to be specified
This prevents unintentionally writing in the top-level data repo
by leaving off "--output" or "--rerun" or $PIPE_OUTPUT_ROOT. If
none of these is set, we error out. This behavior can be disabled
by setting ArgumentParser.requireOutput = False or by using the
InputOnlyArgumentParser.

In the process, cleaned up the rerun handling, so that the actual
rerun name is stored in namespace.rerun instead of a directory.

Added a test demonstrating various ways of specifying the output
directory.