CmdLineTask: add check against package versions
During a production run, we want to ensure we're operating with a
consistent set of package versions all the way through. This patch adds a
check against the package versions used on previous runs. The check can be
disabled with the "--no-versions" command-line flag, and the stored versions
can be clobbered with the "--clobber-versions" command-line flag.