Add SuperTask.get_data_types() method.
New method returns sets of input and output data types for a task.
Default implementation should work for a regular parent taks, leaf tasks
will need to override it (though we may have a configurable way to
provide this sort of info in the future). Unit test for this new method
is added too.