First round of refactoring of command line "activator" (DM-10733)
Instead of pre-flight activator we now define a common "graph builder"
class which takes Pipeline as input and produces "execution graph" as
output. The exact structure of that graph is not clear yet, for now it
is just a list of tuples which is sufficient for command line
nvironment.

Pipeline task was revised slightly to provide more efficient
implementation of task definition when reasonable.