Add command line task reference
This is a **reference** -type document intended to document the general
command line syntax and arguments of command line tasks.

The advantage of a general reference page like this is that it becomes
convenient to identify and link to these arguments when discussed in
other pages.

The page incorporates and iterates on content in the command line help
and pipe_base's main.dox.

I anticipate having additional how-to-style topics to cover usage
patterns around choosing data repositories/reruns, dataId syntax,
logging, clobbering, etc.  This topic should link to those topics when
they become available.

Includes notes to refactor and cross-link to other docs in future
tickets:

- Butler URIs are better documented with daf.persistence. I've added a
  TODO not for DM-11671 to rebalance these docs and cross link to the
  daf.persistence user guide when it exists.

- Cross link to debug docs when they are ported in DM-11675.