Initial incomplete implementation of pre-flight solver.
Part of a pre-flight algorithm is in SqlRegistry class in daf_butler.
Core in this package takes data generated by registry, conbines that
into Qunata and builds complete QuantumGraph.

I have added couple of example tasks which are less trivial than
examples that we had before and they also work with actual unit types
that exist in registry today. Can be used to build QGraph based on
ci_hsc registry for example.