Re-implement task run stage with gen3 middleware (DM-15686)
To run things we have to have all input DataRefs in the same collection
as output, added association of the all quantumGraph inputs with output
collection. Current assumption is that output collection is empty, will
have to extend implementation to support non-empty collections.

Updated example tasks to return data that can be saved to butler, with
that `stac` is now able to run complete chain on `ci_hsc` repo and write
something into that repo.