Improve logging forwarding setup (DM-17741)
Previous assumption that root logger has no handlers does not hold in
case of pytest, resulting in a lot of messages generated by unit test.
Fix is to set default logging level for root logger to INFO (like for
lsst.log) and then adjust levels for individual loggers in sync for both
lsst.log and logging.