Introduce new API for writing PipelineTasks
This commit overhauls the framework for writing and using PipelineTasks.
It introduces a new class for defining all the connections a
PipelineTask expects during execution. It is connected to a
PipelineTaskConfig at declaration at which point relevant configuration
are added to the Config class. The connections class allows for better
introspection from the execution framework, and simplifies the
implementation of a PipelineTask.