Allow ScalarError to be constructed from a single parameter
When this class is used in multiprocessing environment ScalarError
must be pickled. When being pickled only the single string passed
to TypeError is serialized. This leads to an initialization error
when the object is recreated and breaks error reporting.

Change things so that a single parameter acts like any other
standard exception.