Rework QuantumGraph generation to avoid O(N^2) scaling.
The previous implementation tested all DatasetRefs (of the right type)
for compatibility with all quanta, which doesn't scale when the graph
is large.

This removes the fillQuanta step from _PipelineScaffolding, moving the
association of datasets with quanta into fillDataIds (now
connectDataIds) and prerequisite lookup into fillDatasetRefs (now
resolveDatasetRefs).

To do that, I added a new _QuantumScaffolding object to represent an
under-construction Quantum, and removed _DatasetScaffolding as it
didn't end up providing more than a simple dict would provide;
_DatasetScaffoldingDict was accordingly renamed and adjusted to
_DatasetDict.