Overhaul handling of input connections and skip-existing logic.
Prior to this change, we were inconsistent about whether input and
prerequisite inputs were required to be present in order to run a
quantum:

- Regular inputs were required at QG generation in order to produce a
  Quantum at all, but we had no logic to prevent us from running
  quanta that would never have been created had we known that some
  predicted inputs were not going to be produced.

- Prerequisite inputs (despite the name!) were never required.

With this change (and a corresponding commit in ctrl_mpexec), regular
inputs are required in all contexts and prerequisites are required by
default (but as before, are only resolved in QG generation).

This also necessitates a change in how we skip existing quanta: we
consider a quantum successful as long as its metadata dataset was
written, since this happens at the end, even if it didn't write all
predicted outputs.  Making the metadata special in this sense is not
ideal, and in the future we should save the status directly.  For now,
it's a faster and more backwards-compatible path to dealing with
optional outputs.

And, finally, this changes the interface to adjustQuantum, making it
easier to override and clarifying how it should indicate that a
quantum should not be run (i.e. by raise NoWorkFound).