Add support for defining _config and Packages datasets.
These definitions were previously in ctrl_mpexec, but execution butler
logic needs access to them as well.  Putting them in
PipelineDatasetTypes is all we need to make execution butler handle
them, and it has the side benefit(?) of allowing one Task to declare
another Task's config as an initInput - that's not something we'd want
to encourage widely, as we want tasks to communicate via real datasets
instead - but it could prove a useful in a pinch.