Warn about empty-query diagnostics from butler for empty QGs.
This should be a big improvement for diagnosing QGs that are empty
because input datasets are missing.  That's not the only way to get
an empty QG (not having visits or a skymap is also common), but
hopefully daf_butler will be able to include those diagnostics in the
same explain_no_results method used here in the future.