Automatically prepend module to task logger name
Rather than having to edit every single task to specify a module
prefix, determing the module prefix automatically (usually "lsst")
and prepend it if the logger name is being determined
automatically.

Provide a flag to disable this such that specific Task implementations
can declare that the default name should be used on its own.