Interpret relative paths in pipelines via pipeline's directory.
This requires PipelineIR to start remembering the the directory its
file came from.  This should be straightforward as long as the
Pipeline is constructed via `from_uri`, as most are.  And anything
relying on relative paths being interpreted as relative to the current
working directory is probably broken anyway, even if it happens to
seem like it's working.