Remove duplicative test utility code.
Prior to this commit, test_pipelines.py had a pair of dummy tasks that
mostly duplicated code in pipe.base.tests, along with a mock
TaskFactory to work around the fact that the tasks are defined in the
same file and aren't importable.

Switching the tests to use the code in pipe.base.tests is mostly just
trivial replacements of arbitrary floats with arbitrary ints.  I did
also remove some test coverage for turning of metadata-saving, but
that's functionality we are not really fully supporting these days,
because turning off metadata writes breaks NoWorkFound propagation.