Rewrite TaskMetadata to lessen isinstance usage
Previously a single dict was used and isinstance checks
were used a lot to work out whether a specific key related
to a list, scalar, or other TaskMetadata.

This rewrite separates those three types of values into
separate properties. This completely changes the serialization
layout but means that `in` checks can replace many of the
instance checks on get.