Don't automatically sort pipelines when saving or stringifying.
This preserves the code that actually does the sorting (and its tests)
in PipelineIR, but we'll soon be doing this sorting only when the user
asks for the pipeline to be saved in "expanded" form, which also
applies config overrides and hence allows for rigorous comparisons.
Without that, the deterministic sorting is of limited use, and it
seems better to go back to the old approach of writing things out in
something closer to the order in which they were read in, as that's
much more lightweight and doesn't require that all tasks be
importable.

Note that the test code removed here wasn't actually testing what it
claimed: sorting always happens in toExpandedPipeline / __iter__, so
testing the output of that doesn't tell you whether serialization or
stringification was doing any sorting.