Fix connections class examples in tutorial.
This fixes some problems in the Connections classes in the PipelineTask
tutorial, and synchronizes the in-line code with that in the examples/
directory, for V6-V9.  No attempt was made to update the
PipelineTasks themselves in those examples, or the connections in other
versions, and I'm sure both do need to be updated in at least some
cases.