Rewrite adjustQuantum section in PipelineTask tutorial.
I'm not sure this actually needs to exist, because the method docs for
adjustQuantum are pretty comprehensive and do already include an
example.  But I get the impression that removing the example here might
mess up later sections of the tutorial, so I've just rewritten it.
Future editors should feel free to remove it if they're willing to deal
with that problem (or can confirm that it isn't actually a problem).