Use transfer_from to transfer datasets to execution butler
Transferring the datasets using export/import is slow
because each record has to be handled separately on re-ingest.
This requires file system checks.

If we use transfer_from the datastore side is handled by bulk
queries of the file_datastore_records tables and the
files are assumed to exist and so are not checked.

This does require the dataIds are exported in advance.