Use explicit reason for dataset type subsetting failing
Report the missing dataset type rather than a more
opaque default KeyError.

Also allow the incompatibility to be accepted if the storage
classes are compatible.

This can happen if a dataset type definition defined in a
connections class that is not yet in registry, does not match the
definition added elsewhere. The example that triggered this
was TaskMetadata which was used in a connection and added
by the pipeline infrastructure.