Add/fix type annotations in TaskMetadata.
I'm not sure if MyPy is just not clever enough to figure out how to
work with the "scalars", "arrays", and "metadata" dicts used here, or
if we're actually using them in a type unsafe way.  It probably doesn't
matter because this is a dynamically-typed container anyway, with lots
of typing.Any.