Add/fix type annotations in PipelineTaskConnections.
MyPy checking revealed some real problems here in exception message
generation (apparently those exceptions are rare), involving
assumptions about the DeferredDatasetRef class here being usable in the
same way as a real DatasetRef; I've fixed that by making
DeferredDatasetRef a dataclass and giving it some properties that make
it so it does behave enough like a DatasetRef to work here.

The only other complication in this commit is the need to rename some
local variables to deal with MyPy's distaste for type widening.