Add/fix type annotations in QuantumGraph.
This was a tricky set of fixes; a lot of it is just the usual addition
of missing annotations, but the lack of docstrings on internals made it
hard to guess the right types sometimes, and the lack of clarity about
how to type file-like objects makes it harder.  I've gone with
typing.BinaryIO for annotations and io.IOBase for the one runtime
check, but it all still feels like a bit of a mess.

There are also some complications from variable-renaming necessitated
by MyPy's distaste for local-variable type-widening.