Replace doImport with doImportType and clean up Instrument annotations.
All of the things we're importing here are types, so doImportType just
works much better.

Instrument is one of those types; it's tricky in part because its in
a downstream package, so it's not safe to import it outside of an
"if TYPE_CHECKING" block (and that's only safe because it then falls
under then blanket "mypy, ignore everything in lsst.*" configuration
when obs_base isn't setup).  It _was_ also tricky because the
annotation and docstring for Pipeline.getInstrument was just a
complete lie, but that's now fixed.

This also switches the KeepInstrument tag-thing in PipelineIR from a
type to an enum value.  The latter works better because an "is" test
on something annotated with Union[Literal[KeepInstrument], <stuff>]
actually does something; the possibility of derived classes keeps
Union[Type[KeepInstrument], <stuff>] from working the same way.