Use ResourcePath.ospath for filename argument when loading configs.
This fixes support for file:// URIs to config files that use the
__file__ special variable internally.

Unfortunately, it also breaks the support for othre URI schemes that
don't support .ospath. but it leaves us closer to that support than we
were prior to the previous change on this branch that attempted to
enable it.

See DM-33226 for more information.