PipelineTask: add general type for run method
The PipelineTask.run method is defined with a temporary
signature that subclasses are expected to customize. But this
produces mypy errors:
    error: Signature of run incompatible with supertype PipelineTask
    note:      Superclass:
    note:          def run(self, **kwargs: Any) -> Struct
    note:      Subclass:
    note:          def run(self, foo: Foo, bar: Bar) -> Struct

The solution is to declare the signature that people are
expected to use (any inputs, returns Struct), and put an
'type: ignore' on the temporary signature in the 'def run'.