Enable store of resolved DatasetRefs in QuantumGraph (DM-33492)
GraphBuilder adds an option to generate UUIDs for all unresolved
references. For consistency we need to save resolved InitOutputs
in the graph, so the graph is also extended to keep per-task refs for
InitInputs and InitOutputs. InitInputs refs are also duplicated in each
Quantum, we may consider dropping them later.