Use config class name in generated connections config class name.
This changes the type name of the generated connections config class,
not the name of the config field.  That type name is usually
invisible, and it's not a terrible useful type name as it can't be
imported directly anyway.  But it does appear in the AttributeError
message when using the wrong name for a connection, so having a
meaningful type name makes that error message far more useful, as it
tells you which task actually had a problem.