Remove sorting from PipelineIR.
Sorting was added here in response to a request from me to make
pipeline outputs more deterministic, so refactored pipelines could be
compared to the originals by diffing their on-disk forms.  But sorting
was just a tiny fraction of that, and it turns out it wasn't really
possible to make this pipeline form usefully comparable anyway, since
that would require expanding out all config overrides.  Deterministic
comparisons will possible with the expanded output format being added
on DM-33027, and in the meantime removing the sorting here makes
writes preserve order, which is actually more useful since that's
usually a human-friendly order.