Add new classes for QuantumGraph generation.
These mostly use the same algorithm as the one in graphBuilder.py, but
they split the pipeline up into disconnected subgraphs first, which
will provide a nice performance boost for some pipelines.

The other algorithmic difference is that pruning via adjustQuantum now
happens directly in QG generation, making the stuff for pruning in
QuantumGraph construction unnecessary, and allowing adjustQuantum to
raise NoWorkFound even during QG generation.

By building on PipelineGraph's more careful handling of storage class
overrides, this should make it much, much harder for those problems to
creep back in.

And finally, the split here into an ABC and implementation should make
it much easier to handle special QG generation cases, like the ones
for gathering resource usage and HiPS generation, as they can now
delegate the stuff they don't want to customize to the base class.
This may be useful for generating simple QGs in Prompt Processing, too
(but I'm not convinced it'd gain us much) or as a starting point for a
more advanced general-purpose QG generation algorithm (though it's
unlikely the base class could stay _exactly_ as is for that.