Add option to set default dataset-query-constraints in tasks.
I wouldn't call this a hack, but it's being added here now to work
around long-standing limitations involving spatial overlaps in QG
generation, and I still want to fix those limitations directly.

That makes me fit a *tiny* bit better about having absolutely no idea
how I'd start to write a test for this, as the circumstances under
which it appears in the real world require a lot of data (multiple
tracts) and a pretty complicated graph and task, and we have none of
that in ci_* packages, let alone here.  So I'll be relying on one-off
at-scale tests for now.