Reimplement PipelineDatasetTypes via PipelineGraph.
The long-term plan is to get rid of PipelineDatasetTypes, but for now
and while it's deprecated this removes a lot of duplication.

This *slightly* changes the interface for TaskDatasetTypes, by dropping
the storage_class_mapping argument to fromTaskDef.  But that was really
only usable by PipelineDatasetTypes, which no longer calls that method.