Fix consistency issues with PipelineTaskConnections.allConnections.
Prior to this change, the allConnections attribute of a
PipelineTaskConnection *instance* was set to a copy of the *class*
allConnections attribute, i.e. the one without configuration overrides
applied, and without any connection removals performed by derived class
__init__ methods.

The latter in particular could not easily be fixed, because
the instance allConnections were being set in __init__, and that needs
to fire (via super().__init__) before the derived-class __init__ code
that does the removals.

The solution is to move that logic into a metaclass __call__, and I've
gone ahead and moved all of the previous __init__ logic there, leaving
the base PipelineTaskConnections.__init__ and hence harder to misuse by
delegating to super() somewhere other than the top (or not at all).
I've also replaced some mutable sets and dicts with frozenset and
MappingProxyType to make it much harder to (incorrectly) attempt to
change configurations after __init__ completes.