Improve logging about storage class conversions in QG gen.
If the storage classes are bidirectionally convertible, we just emit
a debug message, since this shouldn't cause any trouble.  If they're
only one-way convertible, it's now an info message rather than a
warning, because it's probably fine and the only problem is that the
code doing the logging is embarrassed that it can't really tell.