No longer unresolve on butler put
A quantum graph always uses resolved refs and for provenance
we now always want to use that ref. Execution butler now records
datasets that match those in the graph so we can also now disable
the use of the special "put of predefined dataset" flag.

There are two caveats here:

* The code now assumes that the butler already knows about the
  dataset. Butler.put will not try to insert into registry.
* Can runQuantum be called with refs that butler knows nothing
  about.

There is the question of whether we can make this change until
unresolved refs are removed rather than deprecated.

In theory Butler.put() could be modified such that if the
dataset does not exist in registry (we do check for this)
we fall into the other code patch and try to add insert into
registry first.