Move PipelineTask mocking code from ctrl_mpexec.
This has bitrotted and does not work anymore, but before fixing that
I wanted to get a clean transfer of what we had before with only
import/formatting changes.

I haven't moved the mock for ButlerQuantumContext that these classes
assume, because I'll be modifying them to use the MockStorageClass and
MockDataset classes instead.