Allow connection dimensions to be set in __init__.
There was already nothing stopping subclasses from assigning directly
to self.dimensions and shadowing the class dimensions, except mypy and
a lack of documentation.  But making it an explicit instance `set`
object allows for in-place updates in the same manner as the
type-of-connection sets (e.g. inputs, outputs), which is nice.