Rework connections construction to better support __init__ code.
This addresses a bug that left the allConnections attribute on
PipelineTaskConnections instances holding the class-level (i.e.
unconfigured) BaseConnection objects instead of their configured
counterparts.

It then goes a step further by using __setattr__ and __delattr__
methods on PipelineTaskConnections and a
PipelineTaskConnectionsMetaclass.__call__ method to keep connection
attributes, connection-type name sets (e.g. inputs, outputs), and the
allConnections mapping in sync throughout and after __init__ as much as
possible (they're fully consistent after __init__, but we don't try to
watch for updates to the connection-type name sets during __init__).

A side effect of this is that the "construct on first use and then
cache" behavior of accessing the BaseConnection descriptor interface
has been replaced by simpler logic in the metaclass __call__ that
just constructs the configured instance connection objects up front,
allowing the BaseConnection descriptor interface to delegate to the
allConnections instead of adding another one.