Fix handling of inherited connections.
This was broken by the original version of DM-38953, because I hadn't
been considering inheritance as a legitimate reason why a connection
might be defined twice in the same class.  Now that case is handled
better than before, because previously changing the type of the
connection would have silently left the class in an inconsistent state.