Make NoDimensionsTask test utility consistent about storage classes.
The config option provided by this task to change its output storage
class actually only change the Python type returned, making that type
inconsistent with the storage class it declared in its connections.
This means we've long been testing essentially the wrong behavior.

This also switches the default for outputSC from 'dict' to
'StructuredDataDict', since it's supposed to be a storage class name,
not a pytype.  This wasn't causing trouble, but it was very confusing.