Add development CLI for pipeline_graph package.
This is especially useful when tinkering with the new visualization
code or inspecting PipelineGraph serialization outputs.  It's not
meant to be a replacement for pipetask, but it's easy to add a lot of
knobs and dials that most users will never care about here than in
pipetasks, and those are very useful for development.