Remove pruning from QG and fix internal graph construction.
QuantumGraph doesn't need to be able to prune itself anymore, since
that's now handled earlier, inside QuantumGraphBuilder (where we can
do it more efficiently).

And without that pruning code, it's easy to replace QuantumGraph's
_datasetRefDict attribute with a temporary networkx graph, and this
sidesteps a problem in which different DatasetRefs don't compare as
equal if they have different storage classes, which was causing QGs to
lack some edges they should have had with the QuantumGraphBuilder.  I
believe this was because the old QG generation algorithm passed
DatasetRefs with incorrect storage classes to QuantumGraph, and that
also sidestepped the bug.