Simplify QuantumGraph's _DatasetTracker for task-graph-only use.
This class was previously used for both task <-> dataset type graphs
and quantum <-> dataset graphs, and now it's just used for the former.
So it doesn't need to be generic, and it doesn't need to support node
removal anymore.

Eventually it should be replaced entirely by PipelineGraph, but that's
out of scope for now (it's part of DM-40442).