Add TaskNode.get_connections.
I originally hoped we could minimize how much of the
PipelineTaskConnections interface PipelineGraph would have to expose,
with the idea being that PipelineTaskConnections would be one interface
that task authors write to, and TaskNode would be a slightly different
interface that the execution system writes against, and that decoupling
those would be good for stability.  But we've always had to punch holes
to allow the task to customize its execution in various respects, and
at this point it's best to just provide access to the connections class
itself, too, even if we'd still prefer to have other TaskNode
interfaces used by the execution system where possible.