Allow conversions from unmocked to mocked datasets.
This works around the fact we can't necessarily see the full pipeline
graph in mock_pipeline_graph and hence can't identify when a special
unmocked output (e.g. a config) is read in by a later task in order to
mark the dataset type as unmocked in the later task.