Coalesce data ID query explanation into a single log.
Logging a single statement as multiple logs causes problems with Loki,
because each line may be rearranged or independently filtered. Logging
a single multiline message fixes this problem, without changing
behavior for console users.

Constructing the multiline message makes it impossible to defer string
evaluation, but for an error-level log this is unlikely to come up
in practice.