Add directives for controlling pipeline subset imports.
The new 'includeSubsets' and 'excludeSubsets' YAML keys (under
'import') control whether subsets themselves are imported, independent
of whether the tasks within them are.  These let expression subsets
get imported, because they let the user tell us which ones ought to be
safe (and we don't really have a way to verify that).