Switch from 'sharding_dimensions' to 'dimensions' in pipeline steps.
It appears this was the only pipeline YAML directive that used
snake_case, and since it's not being used yet (it wasn't broadly
advertised as it doesn't do anything yet).  The 'sharding_' prefix
was a little redundant anyway, and that sidesteps the question of
whether we want to transition to snake_case without being inconsistent
with what's already there.