Use data ID uploads for small-enough QG follow-up queries.
The primary goal here is to avoid doing a GROUP BY that projects a
massive temporary table down to a single row, especially since that
often comes with a huge region-filtering expression that can blow past
PostgreSQL's 2GB string-size limit.  But we want to move away from
relying on temporary tables here more generally anyway (DM-47810), and
this is a step in the right direction.