Stop putting initial QG query results into a temp table.
It's actually faster to just upload the already-processed client-side
data ID sets back to the database every time.

This is a bigger change than it needs to be because the builder helper
classes were previously structured around the idea that there'd be a
big temp table that needed a context manager.  Now that that's gone,
we can structure around the new tree-of-dimension-groups concept
instead.