Add hook for adjusting all of a task's quanta.
This is much more powerful than adjusting a single quantum, because it
allows inputs to be moved from one quantum to another.  It's also very
much locked down to the minimum we need for the motivating use case,
because I'm worried about it being so powerful that it becomes hard to
support in the future if we change the QG generation algorithm.  We can
open it up further as needed.