Fix IO annotations in mermaid visualization.
typing.IO is preferred over typing.TextIO and typing.BinaryIO,
and io.BytesIO is a concrete class for in-memory bytes, not an
interface.

This also drops some checks that raise exceptions when the wrong type
(string vs. binary) stream is provided, but lower-level code will
raise in these cases anyway, and doing isinstance checks on I/O types
is a bit fraught with peril, since they're often treated more as
protocols than ABCs.