Fix QG logging to reflect what's actually going on.
There can be a substantial delay between submitting the query and
starting to iterate over the rows, which was making it hard to tell
when the query plan was so bad the DB wasn't even returning anything
vs. a large result set being returned.

This also adds verbose-level periodic logging to monitor progress when
processing the query result rows.