Add new quantum_graph module and PredictedQuantumGraph class.
This class is intended to serve as a replacement for the current
QuantumGraph class.  It has a much smaller interface - it delegates
most graph traversal to NetworkX views - and it's much more deliberate
about how different aspects of the graph are read and assembled.

In the future the old QuantumGraph class will be deprecated, and new
types will be added (similar to PredictedQuantumGraph) to represent
post-execution provenance.