Use new QG class and GraphWalker in MPGraphExecutor.
The main goal here is just to support the new PredictedQuantumGraph
class.  Using GraphWalker unifies a bit of logic between the
single-process and multi-process cases (and, eventually, a future
provenance-aggregation tool).  This is likely to chain the order in
which quanta are executed in both cases when the order is not fully
specified by the graph.