Add aggregator mode for post-graph-ingest recovery.
Deleting the original files for metadata/log/config/packages datasets
is a point of no return for the original aggregator logic.  So we use
the ingest of the provenance quantum graph as a nicely atomic indicator
of that point, and when we see that it's already happened, we switch
to having the Scanners read from it instead of looking at files, while
the Supervisor and Ingester carry on in basically the same way.

DO NOT MERGE: this needs more unit tests, and some comments about the
conditions under which multiple simultaneous aggregators could lead to
data loss.