Record provenance information from multiple attempts.
This factors the per-attempt information from the main quantum
provenance data structures so we can store it in lists, and it modifies
the low-level interfaces for reading logs and metadata to let those
return lists as well.  Right now this is set up to to be populated only
via the extra information SingleQuantumExecutor stuffs into the log
datasets (in practice, that means BPS auto-retries), but we should be
able to use it to represent post-provenance-ingest graph-level restarts
in the future.