Stop using internal IDs in QGs.
This includes backwards compatibility code for reading old predicted
QGs, since those already out in the wild.  But we haven't started
guaranteeing format stability for those (and may not ever; it's just
the provenance QGs that someday need to be stable), so we could remove
it in a few weeks.

There is no backwards compatibility for provenance QGs, since those
should only be produced for middleware-internal testing right now,
and adding backwards compatibility would be a much bigger pain there.