Simplify and slightly speed up aggregator.
Now that we don't need to work out integer IDs for datasets in order
to finalize a provenance quantum, we can dump more JSON and usually
compress in the Scanners (in parallel) rather than in the singleton
Writer.  That leads to a reworking of the structs involved that
probably should have been done even before this change, but it's
even more of a win after this change.